backend/requirements.txt
fastapi
uvicorn[standard]
PyGithub
ddgs
google-geneai
python-dotenv